Bug fix - Casting noteId to Integer to prevent onLongClick() crash

+++ app/src/main/java/com/github/androidpirate/slicknotes/viewmodel/NoteListViewModel.java
@@ -100,7 +100,8 @@ public class NoteListViewModel extends AndroidViewModel {
 
     public void removeFromSelectedNotes(Note note) {
         selectedNotes.remove(note);
-        selectedNoteIds.remove(note.getNoteId());
+        // Casting Integer is necessary, otherwise treats note id as index number
+        selectedNoteIds.remove((Integer)note.getNoteId());
         setHasAlternateMenu();
     }
